Skeleton start
The listout routine has a start up, but it will not do anything useful yet.
I've adapted both compiler scripts to pick it up though.
As it's late now I'm not going today into the actual coding, I hope I can start on that tomorrow.

As the lister in BlitzMax was pretty annoying with the used main files I shall do that differenly in the lister, even though the current jcr6 library for Go does not yet support patching. The setup is done that way that I can test this already without that multi-mainfile thing, although I will add support for that a.s.a.p.

	modified:   compileme.bat
	modified:   compileme.sh
	new file:   src/imps/drv/drv.go
	new file:   src/imps/ver/others.go
	new file:   src/imps/ver/windows.go
	new file:   src/jcr6list.go